requirements.txt
google-api-python-client==2.123.0
pandas==2.2.2
python-dotenv==1.0.1

